DDT-487 updated tests to align with latest version of BKG API
